Extract demo-app's ember-native bootstrap boilerplate into library helpers
demo-app/app/native/setup-ember-native.ts, app/app.js, and app/native/main.ts
each hand-rolled the same non-customizable ember-native wiring (DOM shim
setup, the dev-only DevTools import with its release-build tree-shaking
requirement, the Application buildInstance() override, and Application
creation/config registration) that every consuming app needs verbatim.
Extracted into three new library exports - setupEmberNativeApp(),
NativeApplication, and createNativeApplication() - so demo-app's own files
now only contain what's actually app-specific. Verified via a real
vite build in both production mode (DevTools code confirmed tree-shaken
out) and dev+HMR mode (confirmed present), plus glint/eslint on both
packages.
