+++ ember-native/src/index.ts
@@ -1,5 +1,8 @@
 export { setup } from './setup.ts';
 export { setupInspectorSupport } from './setup-inspector-support.ts';
+export { setupEmberNativeApp } from './setup-app.ts';
+export { createNativeApplication } from './create-native-application.ts';
+export { NativeApplication } from './native-application.ts';
 export * from './components/index.gts';
 export * from './modifiers/index.ts';
 export { default as NativeElementNode } from './dom/native/NativeElementNode.ts';
